disk.CheckUsage: fix floating-point math
Due to accumulated representation errors, CheckUsage() might report
failures in cases where the available space was in fact greater than the
required space, leading to very confusing terminal output.

For example: the default free space ratio of 0.6 cannot be exactly
represented in IEEE-754 floating point. The CLI parsed this as a
float32, which was then converted to a float64 inside the hub and
agents, with a value very slightly greater than 0.6. If a disk failed
the check, we would calculate the required disk space a second time,
using a method that was subtly different from our original check. The
combined discrepancies led to a less-than-sane result.

Two-part fix:
- calculate required disk space only once, and use integer comparison to
  determine whether the check succeeds or fails, to prevent failure
  reports where avail > required
- use float64 from the beginning, during command-line parsing, to avoid
  exacerbating representational error later down the line

To assist with any future debugging, gplog.Debug() output has been added
during disk space checks.

+++ cli/commanders/checks.go
@@ -14,7 +14,7 @@ import (
 	"github.com/greenplum-db/gpupgrade/utils/disk"
 )
 
-func RunPreChecks(client idl.CliToHubClient, ratio float32) error {
+func RunPreChecks(client idl.CliToHubClient, ratio float64) error {
 	//TODO: when do we check this?  It requires the source cluster to be up.
 	//err := CheckVersion(client)
 	//if err != nil {
@@ -115,7 +115,7 @@ func (t tableRows) Swap(i, j int) {
 	t[i], t[j] = t[j], t[i]
 }
 
-func CheckDiskSpace(client idl.CliToHubClient, ratio float32) (err error) {
+func CheckDiskSpace(client idl.CliToHubClient, ratio float64) (err error) {
 	s := Substep("Checking disk space...")
 	defer s.Finish(&err)
